+++ src/Pim/Product/ProductListFilter.php
@@ -49,6 +49,9 @@ final class ProductListFilter implements ArrayableInterface
     /** @var null|string */
     private $updatedAfter;
 
+    /** @var int[] */
+    private $companyIds = [];
+
     /**
      * @param string $productCode
      *
@@ -141,6 +144,17 @@ final class ProductListFilter implements ArrayableInterface
     }
 
     /**
+     * @param int[] $companyIds
+     * @return ProductListFilter
+     */
+    public function byCompanyIds(array $companyIds): self
+    {
+        $this->companyIds = $companyIds;
+
+        return $this;
+    }
+
+    /**
      * @internal
      */
     public function toArray(): array
@@ -184,6 +198,10 @@ final class ProductListFilter implements ArrayableInterface
 
         if (\count($this->supplierReferences) > 0) {
             $filters['supplier_ref'] = $this->supplierReferences;
+        }
+
+        if (\count($this->companyIds) > 0) {
+            $filters['company_ids'] = $this->companyIds;
         }
 
         return $filters;